addressing process_ctd
